// ambari-web/latest/src/screens/Metrics/type.ts
export type WidgetInfo = {
    id: number;
    metrics: string;
    author: string;
    description: string;
    scope: string;
    properties: string;
    tag: any;
    widget_name: string;
    widget_type: string;
    time_created: number;
    display_name: string;
    values: string;
    cluster_name: string;
  }